Add EventDispatcher

+++ core/src/com/scramble_like/game/component/TestComponent.java
@@ -11,9 +11,9 @@ public class TestComponent extends Component
     @Override
     public void Update(double DeltaTime)
     {
-        this.getOwner().getTransform().getRotation().x += 10 * DeltaTime;
-        this.getOwner().getTransform().getScale().x += 0.5f * DeltaTime * ScaleDirection;
-        this.getOwner().getTransform().getScale().y += 0.5f * DeltaTime * ScaleDirection;
+        this.getOwner().getTransform().getRotation().x += (float) (10 * DeltaTime);
+        this.getOwner().getTransform().getScale().x += (float) (0.5f * DeltaTime * ScaleDirection);
+        this.getOwner().getTransform().getScale().y += (float) (0.5f * DeltaTime * ScaleDirection);
         if (this.getOwner().getTransform().getScale().x > 1.5) { ScaleDirection = -1; }
         if (this.getOwner().getTransform().getScale().x < 0.5) { ScaleDirection = 1; }
     }
